@@ -1,20 +1,15 @@
-name: server
-description: A server app using the shelf package and Docker.
-version: 1.0.0
-# homepage: https://www.example.com
+name: flow_server
+description: An new Dart Frog application
+version: 1.0.0+1
+publish_to: none
 
 environment:
-  sdk: '>=2.18.1 <3.0.0'
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
-  args: ^2.4.0
-  collection: ^1.17.0
-  dotenv: ^4.1.0
-  shelf: ^1.4.0
-  shelf_router: ^1.0.0
-  sqlite3: ^1.10.1
+  dart_frog: ^0.3.0
 
 dev_dependencies:
-  http: ^0.13.0
-  lints: ^2.0.1
-  test: ^1.24.0
+  mocktail: ^0.3.0
+  test: ^1.19.2
+  very_good_analysis: ^4.0.0
