@@ -15,7 +15,7 @@ class Appointment with _$Appointment implements EventItem {
     @Default('') String name,
     @Default('') String description,
     @Default('') String location,
-    required int eventId,
+    int? eventId,
     @Default(EventStatus.confirmed) EventStatus status,
     @DateTimeConverter() DateTime? start,
     @DateTimeConverter() DateTime? end,
@@ -26,7 +26,7 @@ class Appointment with _$Appointment implements EventItem {
     @Default('') String name,
     @Default('') String description,
     @Default('') String location,
-    required int eventId,
+    int? eventId,
     @Default(EventStatus.confirmed) EventStatus status,
     @DateTimeConverter() DateTime? start,
     @DateTimeConverter() DateTime? end,
@@ -43,7 +43,7 @@ class Appointment with _$Appointment implements EventItem {
     @Default('') String name,
     @Default('') String description,
     @Default('') String location,
-    required int eventId,
+    int? eventId,
     @Default(EventStatus.confirmed) EventStatus status,
     @DateTimeConverter() DateTime? start,
     @DateTimeConverter() DateTime? end,
