@@ -4,11 +4,14 @@ import 'package:shared/services/source.dart';
 
 import 'appointment/model.dart';
 import 'model.dart';
+import 'moment/model.dart';
 
 abstract class EventService extends ModelService {
   FutureOr<Event?> getEvent(int id);
   FutureOr<Event?> getEventByAppointment(Appointment appointment) =>
-      getEvent(appointment.eventId);
+      appointment.eventId == null ? null : getEvent(appointment.eventId!);
+  FutureOr<Event?> getEventByMoment(Moment moment) =>
+      moment.eventId == null ? null : getEvent(moment.eventId!);
 
   FutureOr<List<Event>> getEvents({
     int? groupId,
